refactor satisfaction / fulfillment logic into origin/destination library crate pairs
