+++ happs/observation/zomes/satisfaction/code/Cargo.toml
@@ -14,14 +14,12 @@ hdk_proc_macros = "=0.0.40-alpha1"
 holochain_wasm_utils = "=0.0.40-alpha1"
 holochain_json_derive = "=0.0.17"
 
-hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
-vf_core = { path = "../../../../../lib/vf_core" }
 hc_zome_rea_economic_event_storage_consts = { path = "../../../../../lib/rea_economic_event/storage_consts" }
 hc_zome_rea_satisfaction_storage_consts = { path = "../../../../../lib/rea_satisfaction/storage_consts" }
 hc_zome_rea_satisfaction_storage = { path = "../../../../../lib/rea_satisfaction/storage" }
 hc_zome_rea_satisfaction_defs = { path = "../../../../../lib/rea_satisfaction/defs" }
 hc_zome_rea_satisfaction_rpc = { path = "../../../../../lib/rea_satisfaction/rpc" }
-hc_zome_rea_satisfaction_lib = { path = "../../../../../lib/rea_satisfaction/lib" }
+hc_zome_rea_satisfaction_lib_destination = { path = "../../../../../lib/rea_satisfaction/lib_destination" }
 
 [lib]
 path = "src/lib.rs"
