+++ happs/planning/zomes/fulfillment/code/Cargo.toml
@@ -14,7 +14,7 @@ hdk_proc_macros = "0.0.40-alpha1"
 
 hc_zome_rea_fulfillment_defs = { path = "../../../../../lib/rea_fulfillment/defs" }
 hc_zome_rea_fulfillment_rpc = { path = "../../../../../lib/rea_fulfillment/rpc" }
-hc_zome_rea_fulfillment_lib = { path = "../../../../../lib/rea_fulfillment/lib" }
+hc_zome_rea_fulfillment_lib_origin = { path = "../../../../../lib/rea_fulfillment/lib_origin" }
 
 [lib]
 path = "src/lib.rs"
