+++ happs/planning/zomes/fulfillment/code/src/lib.rs
@@ -17,7 +17,7 @@ use hdk_proc_macros::zome;
 
 use hc_zome_rea_fulfillment_defs::{ entry_def, base_entry_def };
 use hc_zome_rea_fulfillment_rpc::*;
-use hc_zome_rea_fulfillment_lib::*;
+use hc_zome_rea_fulfillment_lib_origin::*;
 
 // Zome entry type wrappers
 #[zome]
