+++ lib/rea_satisfaction/lib_destination/src/lib.rs
@@ -1,13 +1,16 @@
 /**
- * Handling for `Fulfillment` related behaviours as they relate to `EconomicEvent`s
+ * Holo-REA satisfaction zome library API
+ *
+ * Contains helper methods that can be used to manipulate `Satisfaction` data
+ * structures in either the local Holochain zome, or a separate DNA-local zome.
+ *
+ * Contains functionality for the "destination" side of an "indirect remote index" pair
+ * (@see `hdk_graph_helpers` README).
+ *
+ * @package Holo-REA
  */
+use hdk::prelude::*;
 
-use hdk::{
-    holochain_json_api::{json::JsonString, error::JsonError},
-    error::ZomeApiResult,
-    error::ZomeApiError,
-};
-use holochain_json_derive::{ DefaultJson };
 use hdk_graph_helpers::{
     records::{
         create_record,
@@ -16,30 +19,16 @@ use hdk_graph_helpers::{
         delete_record,
     },
     local_indexes::{
-        create_direct_index,
         query_direct_index_with_foreign_key,
+        create_direct_index,
     },
 };
 
-use vf_core::type_aliases::{
-    EventAddress,
-    SatisfactionAddress,
-};
-use hc_zome_rea_economic_event_storage_consts::{
-    EVENT_SATISFIES_LINK_TYPE,
-    EVENT_SATISFIES_LINK_TAG,
-};
+use hc_zome_rea_economic_event_storage_consts::{EVENT_SATISFIES_LINK_TYPE, EVENT_SATISFIES_LINK_TAG};
 use hc_zome_rea_satisfaction_storage_consts::*;
-
 use hc_zome_rea_satisfaction_storage::Entry;
 use hc_zome_rea_satisfaction_rpc::*;
 use hc_zome_rea_satisfaction_lib::construct_response;
-
-#[derive(Serialize, Deserialize, Debug, DefaultJson, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct QueryParams {
-    satisfied_by: Option<EventAddress>,
-}
 
 pub fn receive_create_satisfaction(satisfaction: CreateRequest) -> ZomeApiResult<ResponseData> {
     handle_create_satisfaction(&satisfaction)
